Look up keys in db file.
When a key's value is requested Eza will iterate through
each line in the DB file until it has the latest entry
for the given key.

This approach is inefficient and intended to pave the way
for a system where the file scan only happens when Eza
starts up and loads the database.

Closes #1