Change crash test to use missing wal file.
The previous test for a crashed/invalid WAL file required
leaving hard-coded data in the repo. This was both messy
and made it hard to account for changes in the data format.

The new approach deletes the WAL outright after performing a
transaction. The test now ensures that the keys from before the
delete are missing while keys created after the delete are
still fine.