Use persistent primary keys.
This change adds support for locking a key in the database
before performing an operation on it. This is done in order
to support persisting auto incrementing primary keys to disk.

This change also fixes the get_row() implementation by doing
a proper check for the primary key for each k/v pair.