Add function to update a row given its id.
Columns for an existing row can now be updated or added.
This function works similarly to `insert_row` except that it
also takes the id for the row that needs to be updated.

This change also adds a check to ensure that only the most
recently inserted (and committed) column value is fetched in
get_row(). There is also a function to extract the transaction id
for a given row entry's key.